Support for elements meta-annotated with ProtectedResourceId
