+++ security/src/main/java/org/craftercms/commons/security/permissions/annotations/HasPermissionAnnotationHandler.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2007-2022 Crafter Software Corporation. All Rights Reserved.
+ * Copyright (C) 2007-2023 Crafter Software Corporation. All Rights Reserved.
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 3 as published by
@@ -15,12 +15,6 @@
  */
 package org.craftercms.commons.security.permissions.annotations;
 
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Method;
-import java.lang.reflect.Parameter;
-import java.util.HashMap;
-import java.util.Map;
-
 import org.apache.commons.lang3.StringUtils;
 import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Around;
@@ -29,11 +23,12 @@ import org.craftercms.commons.aop.AopUtils;
 import org.craftercms.commons.http.RequestContext;
 import org.craftercms.commons.i10n.I10nLogger;
 import org.craftercms.commons.security.exception.ActionDeniedException;
-import org.craftercms.commons.security.exception.PermissionException;
 import org.craftercms.commons.security.permissions.PermissionEvaluator;
-import org.springframework.beans.factory.annotation.Required;
-import org.springframework.core.annotation.AnnotationUtils;
 import org.springframework.core.annotation.Order;
+
+import java.beans.ConstructorProperties;
+import java.lang.reflect.Method;
+import java.util.Map;
 
 /**
  * Aspect that handles {@link org.craftercms.commons.security.permissions.annotations.HasPermission} annotations,
@@ -43,7 +38,7 @@ import org.springframework.core.annotation.Order;
  */
 @Aspect
 @Order(-1)
-public class HasPermissionAnnotationHandler {
+public class HasPermissionAnnotationHandler extends AbstractPermissionAnnotationHandler {
 
     private static final String TOKEN_PARAMETER = "token";
 
@@ -53,24 +48,15 @@ public class HasPermissionAnnotationHandler {
     private static final String LOG_KEY_METHOD_INT = "security.permission.methodIntercepted";
     private static final String LOG_KEY_METHOD_INT_NO_SEC_OBJ = "security.permission.methodInterceptedNoSecObject";
 
-    private static final String ERROR_KEY_EVALUATOR_NOT_FOUND = "security.permission.evaluatorNotFound";
-    private static final String ERROR_KEY_EVALUATION_FAILED = "security.permission.evaluationFailed";
-
-    protected Map<Class<?>, PermissionEvaluator<?, ?>> permissionEvaluators;
-
-    protected String managementToken;
-
-    @Required
-    public void setPermissionEvaluators(Map<Class<?>, PermissionEvaluator<?, ?>> permissionEvaluators) {
-        this.permissionEvaluators = permissionEvaluators;
-    }
-
     /**
-     * Sets the management token to be validated in case a {@link HasPermission} annotation
+     * Management token to be validated in case a {@link HasPermission} annotation
      * has been configured to accept a management token.
-     * @param managementToken the management token value
      */
-    public void setManagementToken(final String managementToken) {
+    protected final String managementToken;
+
+    @ConstructorProperties({"permissionEvaluators", "managementToken"})
+    public HasPermissionAnnotationHandler(Map<Class<?>, PermissionEvaluator<?, ?>> permissionEvaluators, String managementToken) {
+        super(permissionEvaluators);
         this.managementToken = managementToken;
     }
 
@@ -79,10 +65,10 @@ public class HasPermissionAnnotationHandler {
             "@annotation(org.craftercms.commons.security.permissions.annotations.HasPermission)")
     public Object checkPermissions(ProceedingJoinPoint pjp) throws Throwable {
         Method method = AopUtils.getActualMethod(pjp);
-        HasPermission hasPermission = getHasPermissionAnnotation(method, pjp);
-        Object securedResource = getAnnotatedProtectedResource(method, pjp);
+        HasPermission hasPermission = getHasPermissionAnnotation(method, pjp, HasPermission.class);
+        Object securedResource = getAnnotatedProtectedResource(method, pjp.getArgs());
         if (securedResource == null) {
-            securedResource = getAnnotatedProtectedResourceIds(method, pjp);
+            securedResource = getAnnotatedProtectedResourceIds(method, pjp.getArgs());
         }
 
         if (securedResource != null) {
@@ -122,78 +108,6 @@ public class HasPermissionAnnotationHandler {
         String token = requestContext.getRequest().getParameter(TOKEN_PARAMETER);
 
         return StringUtils.equals(token, managementToken);
-    }
-
-    /**
-     * Checks the permissions to perform the action configured in the {@link HasPermission} to the securedResource (if any)
-     *
-     * @param method          the {@link Method} to secure
-     * @param hasPermission   the {@link HasPermission} annotation
-     * @param securedResource the securedResource, if any
-     * @return true if the action is allowed, false otherwise
-     */
-    @SuppressWarnings("unchecked")
-    protected boolean checkPermissions(final Method method, final HasPermission hasPermission, final Object securedResource) {
-        Class<?> type = hasPermission.type();
-        String action = hasPermission.action();
-        PermissionEvaluator permissionEvaluator = permissionEvaluators.get(type);
-
-        if (permissionEvaluator == null) {
-            throw new PermissionException(ERROR_KEY_EVALUATOR_NOT_FOUND, type);
-        }
-
-        try {
-            return permissionEvaluator.isAllowed(securedResource, action);
-        } catch (PermissionException e) {
-            throw new PermissionException(ERROR_KEY_EVALUATION_FAILED, e);
-        }
-    }
-
-    protected HasPermission getHasPermissionAnnotation(Method method, ProceedingJoinPoint pjp) {
-        HasPermission hasPermission = method.getAnnotation(HasPermission.class);
-
-        if (hasPermission == null) {
-            Class<?> targetClass = pjp.getTarget().getClass();
-            hasPermission = targetClass.getAnnotation(HasPermission.class);
-        }
-
-        return hasPermission;
-    }
-
-    protected Object getAnnotatedProtectedResource(Method method, ProceedingJoinPoint pjp) {
-        Annotation[][] paramAnnotations = method.getParameterAnnotations();
-        Object[] params = pjp.getArgs();
-
-        for (int i = 0; i < paramAnnotations.length; i++) {
-            for (Annotation a : paramAnnotations[i]) {
-                if (a instanceof ProtectedResource) {
-                    return params[i];
-                }
-            }
-        }
-
-        return null;
-    }
-
-    protected Map<String, Object> getAnnotatedProtectedResourceIds(Method method, ProceedingJoinPoint pjp) {
-        Parameter[] methodParameters = method.getParameters();
-        Object[] paramValues = pjp.getArgs();
-        Map<String, Object> resourceIds = null;
-
-        for (int i = 0; i < methodParameters.length; i++) {
-            ProtectedResourceId resourceIdAnnotation = AnnotationUtils.findAnnotation(methodParameters[i], ProtectedResourceId.class);
-            if (resourceIdAnnotation != null) {
-                String idName = resourceIdAnnotation.value();
-
-                if (resourceIds == null) {
-                    resourceIds = new HashMap<>();
-                }
-
-                resourceIds.put(idName, paramValues[i]);
-            }
-        }
-
-        return resourceIds;
     }
 
 }
